merge

+++ services/subnet_manager/src/main/java/com/futurewei/alcor/subnet/controller/SubnetController.java
@@ -296,12 +296,15 @@ public class SubnetController {
 //            if (macState != null) {
 //                inSubnetEntity.setGatewayMacAddress(macState.getMacAddress());
 //            }
+            this.subnetDatabaseService.addSubnet(inSubnetEntity);
+
             if (gatewayIpIsInAllocatedRange) {
                 PortEntity portEntity = this.subnetService.constructPortEntity(portId, vpcId, subnetId, ipResponse.getIp(), ConstantsConfig.DeviceOwner);
                 GatewayPortDetail gatewayPortDetail = this.subnetToPortManagerService.createGatewayPort(projectId, portEntity);
 
                 inSubnetEntity.setGatewayIp(ipResponse.getIp());
                 inSubnetEntity.setGatewayPortDetail(gatewayPortDetail);
+                inSubnetEntity.setGatewayPortId(gatewayPortDetail.getGatewayPortId());
             } else {
                 String gatewayIP = SubnetManagementUtil.setGatewayIpValue(gatewayIp, cidr);
                 if (gatewayIp != null) {
@@ -310,8 +313,10 @@ public class SubnetController {
 
                     inSubnetEntity.setGatewayIp(gatewayIP);
                     inSubnetEntity.setGatewayPortDetail(gatewayPortDetail);
+                    inSubnetEntity.setGatewayPortId(gatewayPortDetail.getGatewayPortId());
                 }
             }
+
             if (ipResponse != null && ipResponse.getIpVersion() == 4) {
                 inSubnetEntity.setIpV4RangeId(ipResponse.getRangeId());
             }else if (ipResponse != null &&  ipResponse.getIpVersion() == 6) {
@@ -427,7 +432,18 @@ public class SubnetController {
             String newGatewayIp = inSubnetWebResponseObject.getGatewayIp();
             if (newGatewayIp == null) {// disable gatewayIP
                 subnetEntity.setGatewayIp(newGatewayIp);
+
+                // delete old gateway port
+                if (oldGatewayIp != null) {
+                    this.subnetToPortManagerService.deleteGatewayPort(projectId,oldPortId);
+                }
+
             } else if(!newGatewayIp.equals(oldGatewayIp)){
+
+                String routerId = subnetEntity.getAttachedRouterId();
+                if (routerId != null) {
+                    throw new CanNotUpdateGatewayPort();
+                }
 
                 // check if updated gatewayIp is valid
                 boolean gatewayIpIsValid = SubnetManagementUtil.checkGatewayIpInputSupported(newGatewayIp, subnetEntity.getCidr());
@@ -441,9 +457,12 @@ public class SubnetController {
                     GatewayPortDetail gatewayPortDetail = this.subnetToPortManagerService.createGatewayPort(projectId, portEntity);
                     subnetEntity.setGatewayPortDetail(gatewayPortDetail);
                     subnetEntity.setGatewayIp(newGatewayIp);
+                    subnetEntity.setGatewayPortId(gatewayPortDetail.getGatewayPortId()); // -tem
 
                     // delete port with old gateway port IP & port Id
-                    this.subnetToPortManagerService.deleteGatewayPort(projectId,oldPortId);
+                    if (oldGatewayIp != null) {
+                        this.subnetToPortManagerService.deleteGatewayPort(projectId,oldPortId);
+                    }
                 } else {
                     throw new GatewayIpUnsupported();
                 }
@@ -457,12 +476,11 @@ public class SubnetController {
                 subnetEntity.setRevisionNumber(revisionNumber + 1);
             }
 
+            // update subnet routing rule in route manager
+            this.subnetService.updateSubnetRoutingRuleInRM(projectId, subnetId, subnetEntity);
 
             this.subnetDatabaseService.addSubnet(subnetEntity);
             subnetEntity = this.subnetDatabaseService.getBySubnetId(subnetId);
-
-            // update subnet routing rule in route manager
-            this.subnetService.updateSubnetRoutingRuleInRM(projectId, subnetId, subnetEntity);
 
         } catch (ParameterNullOrEmptyException e) {
             logger.error(e.getMessage());
@@ -505,10 +523,10 @@ public class SubnetController {
             } else {
                 rangeId = ipV6RangeId;
             }
-            Boolean checkIfAnyPortInSubnet = this.subnetService.checkIfAnyPortInSubnet(rangeId);
-            if (checkIfAnyPortInSubnet) {
-                throw new HavePortInSubnet();
-            }
+//            Boolean checkIfAnyPortInSubnet = this.subnetService.checkIfAnyPortInSubnet(rangeId);
+//            if (checkIfAnyPortInSubnet) {
+//                throw new HavePortInSubnet();
+//            }
 
             // check if subnet bind any routes
             Boolean checkIfSubnetBindAnyRoutes = this.subnetService.checkIfSubnetBindAnyRoutes(subnetEntity);
@@ -516,19 +534,19 @@ public class SubnetController {
                 throw new SubnetBindRoutes();
             }
 
-            this.subnetDatabaseService.deleteSubnet(subnetId);
-
             // delete subnet id in vpc
             this.subnetService.deleteSubnetIdInVpc(subnetId, projectId, subnetEntity.getVpcId());
 
             // delete subnet routing rule in route manager
             this.subnetService.deleteSubnetRoutingRuleInRM(projectId, subnetId);
 
-            // delete gateway port in port manager
+            // TODO: delete gateway port in port manager. Temporary solution, need PM fix issue
             GatewayPortDetail gatewayPortDetail = subnetEntity.getGatewayPortDetail();
             if (gatewayPortDetail != null) {
                 this.subnetToPortManagerService.deleteGatewayPort(projectId, gatewayPortDetail.getGatewayPortId());
             }
+
+            this.subnetDatabaseService.deleteSubnet(subnetId);
 
         } catch (ParameterNullOrEmptyException | HavePortInSubnet | SubnetBindRoutes e) {
             logger.error(e.getMessage());
